refactor: use shared watcher in bpa-server and reorganize static routes
